First step to prepare this project to run jasmine tests.
Added django-jasmine and selenium to requirements.txt
Added jasmine_tests.py and jasmine_tests folder with a first test
that does nothing. All this can be run with the Chrome webdriver
already but I guess this will not work on travis-ci. Therefore
I will try to get this working with the PhantomJS webdriver.

Also added a manage.py. Does not fully work, yet. The goal is to
be able to call ./manage.py runserver and browse to /jasmine/ when
developing in order to get more verbose error output in case that
any jasmine tests fail.